requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
bcrypt
pyttsx3